requirements.txt
anyio==4.6.2.post1
asgiref==3.9.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
Flask==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.1
python-decouple==3.8
python-telegram-bot==21.7
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
urllib3==2.2.3
Werkzeug==3.1.3
